misc/misc400-voices/requirements.txt
click==6.6
Flask==0.11.1
gTTS==1.1.6
gTTS-token==1.1.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
peewee==2.8.5
pydub==0.16.5
python-Levenshtein==0.12.0
requests==2.11.1
SpeechRecognition==3.4.6
Werkzeug==0.11.11
wheel==0.24.0
